Moving Async collection methods to the Async namespace
