+++ docs/articles/collections.md
@@ -1,3 +1,11 @@
 # Collections
 
-TODO
+## Dictionaries, Sets, Queues, Stacks
+
+TODO: Sets, Queues, Stacks - dedicated extension methods needed?
+
+## IEnumerable<T>
+
+## IAsyncEnumerable<T>
+
+
