+++ src/RustyOptions.Tests/OptionAsyncTests.cs
@@ -6,6 +6,66 @@ namespace RustyOptions.Tests;
 
 public sealed class OptionAsyncTests
 {
+    [Fact]
+    public async Task CanGetValuesAsync()
+    {
+        int count = 0;
+        await foreach (var x in EnumerateFilteredAsync(11, i => (i & 1) == 0).ValuesAsync())
+        {
+            Assert.True((x & 1) == 0);
+            count++;
+        }
+
+        Assert.Equal(6, count);
+    }
+
+    [Fact]
+    public async Task CanGetValuesWithCancelAsync()
+    {
+        using var cts = new CancellationTokenSource();
+
+        int count = 0;
+        await foreach (var x in EnumerateFilteredAsync(11, i => (i & 1) == 0, cts.Token).ValuesAsync(cts.Token))
+        {
+            Assert.True((x & 1) == 0);
+            count++;
+
+            if (count > 2)
+            {
+                cts.Cancel();
+            }
+        }
+
+        Assert.Equal(3, count);
+    }
+
+    [Fact]
+    public async Task CanGetFirstOrNoneAsync()
+    {
+        Assert.Equal(None<int>(), await EnumerateAsync(0).FirstOrNoneAsync());
+        Assert.Equal(Some(0), await EnumerateAsync(11).FirstOrNoneAsync());
+        Assert.Equal(Some(2), await EnumerateAsync(11).FirstOrNoneAsync(x => x > 0 && (x & 1) == 0));
+        Assert.Equal(None<int>(), await EnumerateAsync(11).FirstOrNoneAsync(x => x > 500));
+    }
+
+    [Fact]
+    public async Task CanGetFirstOrNoneWithCancelAsync()
+    {
+        using var cts = new CancellationTokenSource();
+
+        Assert.Equal(None<int>(), await EnumerateAsync(0, cts.Token).FirstOrNoneAsync(cts.Token));
+        Assert.Equal(Some(0), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(cts.Token));
+        Assert.Equal(Some(2), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(x => x > 0 && (x & 1) == 0, cts.Token));
+        Assert.Equal(None<int>(), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(x => x > 500, cts.Token));
+
+        cts.Cancel();
+
+        Assert.Equal(None<int>(), await EnumerateAsync(0, cts.Token).FirstOrNoneAsync(cts.Token));
+        Assert.Equal(None<int>(), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(cts.Token));
+        Assert.Equal(None<int>(), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(x => x > 0 && (x & 1) == 0, cts.Token));
+        Assert.Equal(None<int>(), await EnumerateAsync(11, cts.Token).FirstOrNoneAsync(x => x > 500, cts.Token));
+    }
+
     [Theory]
     [MemberData(nameof(GetMapAsyncValues))]
     public async Task CanMapAsync(object source, object mapper, Option<long> expected)
@@ -293,5 +353,33 @@ public sealed class OptionAsyncTests
         yield return new object[] { valTaskSrcNone, taskMapper, expectedNone };
         yield return new object[] { taskSrcNone, valTaskMapper, expectedNone };
         yield return new object[] { valTaskSrcNone, valTaskMapper, expectedNone };
+    }
+
+
+    /// <summary>
+    /// Generates an IAsyncEnumerable that counts up to, but not including, the given
+    /// <paramref name="exclusiveMax"/>
+    /// </summary>
+    private static async IAsyncEnumerable<int> EnumerateAsync(int exclusiveMax, [EnumeratorCancellation] CancellationToken ct = default)
+    {
+        for (int i = 0; !ct.IsCancellationRequested && i < exclusiveMax; i++)
+        {
+            await Task.Yield();
+            yield return i;
+        }
+    }
+
+    /// <summary>
+    /// Generates an IAsyncEnumerable that counts up to, but not including, the given
+    /// <paramref name="exclusiveMax"/> - returning Some for numbers for which the predicate returns true
+    /// and None otherwise.
+    /// </summary>
+    private static async IAsyncEnumerable<Option<int>> EnumerateFilteredAsync(int exclusiveMax, Func<int, bool> predicate, [EnumeratorCancellation] CancellationToken ct = default)
+    {
+        for (int i = 0; !ct.IsCancellationRequested && i < exclusiveMax; i++)
+        {
+            await Task.Yield();
+            yield return predicate(i) ? Some(i) : None<int>();
+        }
     }
 }
