+++ src/RustyOptions/Async/OptionAsyncExtensions.cs
@@ -9,6 +9,10 @@ namespace RustyOptions.Async;
 /// </summary>
 public static class OptionAsyncExtensions
 {
+    // NOTE: Due to a bug in coverlet.collector, certain lines in methods involving IAsyncEnumerable
+    // will show as partially-covered in code-coverage tools, even when they are fully-covered.
+    // https://github.com/coverlet-coverage/coverlet/issues/1104#issuecomment-1005332269
+
     /// <summary>
     /// If the option has a value, passes that option to the mapper function and asynchronously returns that value
     /// as a <c>Some</c>.
@@ -621,5 +625,74 @@ public static class OptionAsyncExtensions
         return opt.IsNone
             ? await elseFn().ConfigureAwait(false)
             : opt;
+    }
+
+    /// <summary>
+    /// Returns the first element of an asynchronous sequence if such exists.
+    /// </summary>
+    /// <param name="self">The sequence to return the first element from.</param>
+    /// <param name="ct">A CancellationToken that will interrupt async iteration.</param>
+    /// <returns>An <see cref="Option{T}"/> instance containing the first element if present.</returns>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="self"/> is null.</exception>
+    public static async ValueTask<Option<T>> FirstOrNoneAsync<T>(this IAsyncEnumerable<T> self, CancellationToken ct = default)
+        where T : notnull
+    {
+        ThrowIfNull(self);
+
+        await using var enumerator = self.GetAsyncEnumerator(ct);
+        if (await enumerator.MoveNextAsync().ConfigureAwait(false))
+        {
+            return Some(enumerator.Current);
+        }
+
+        return default;
+    }
+
+    /// <summary>
+    /// Returns the first element of a sequence, satisfying a specified predicate, 
+    /// if such exists.
+    /// </summary>
+    /// <param name="self">The sequence to return the first element from.</param>
+    /// <param name="predicate">The predicate to filter by.</param>
+    /// <param name="ct">A CancellationToken that will interrupt async iteration.</param>
+    /// <returns>An <see cref="Option{T}"/> instance containing the first matching element, if present.</returns>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="self"/> or <paramref name="predicate"/> is null.</exception>
+    public static async ValueTask<Option<T>> FirstOrNoneAsync<T>(this IAsyncEnumerable<T> self, Func<T, bool> predicate, CancellationToken ct = default)
+        where T : notnull
+    {
+        ThrowIfNull(self);
+        ThrowIfNull(predicate);
+
+        await foreach (var item in self.WithCancellation(ct))
+        {
+            if (predicate(item))
+            {
+                return Some(item);
+            }
+        }
+
+        return default;
+    }
+
+    /// <summary>
+    /// Flattens an asynchronous sequence of <see cref="Option{T}"/> into a sequence containing all inner values.
+    /// Empty elements are discarded.
+    /// </summary>
+    /// <param name="self">The sequence of options.</param>
+    /// <param name="ct">A CancellationToken that will interrupt async iteration.</param>
+    /// <returns>A flattened sequence of values.</returns>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="self"/> is null.</exception>
+    public static async IAsyncEnumerable<T> ValuesAsync<T>(this IAsyncEnumerable<Option<T>> self, [EnumeratorCancellation] CancellationToken ct = default)
+        where T : notnull
+    {
+        ThrowIfNull(self);
+
+        await foreach (var option in self.WithCancellation(ct))
+        {
+            if (option.IsSome(out var value))
+            {
+                yield return value;
+            }
+        }
     }
 }
