+++ src/RustyOptions/Async/ResultAsyncExtensions.cs
@@ -730,4 +730,48 @@ public static class ResultAsyncExtensions
 
         return Result.Ok<T, T2Err>(result.Unwrap());
     }
+
+    /// <summary>
+    /// Flattens an asynchronous sequence of <see cref="Result{T, TErr}"/> into a sequence containing all inner values.
+    /// Error results are discarded.
+    /// </summary>
+    /// <param name="self">The asynchronsous sequence of results.</param>
+    /// <param name="ct">A CancellationToken that will interrupt async iteration.</param>
+    /// <returns>A flattened sequence of values.</returns>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="self"/> is null.</exception>
+    public static async IAsyncEnumerable<T> ValuesAsync<T, TErr>(this IAsyncEnumerable<Result<T, TErr>> self, [EnumeratorCancellation] CancellationToken ct = default)
+        where T : notnull
+    {
+        ThrowIfNull(self);
+
+        await foreach (var result in self.WithCancellation(ct))
+        {
+            if (result.IsOk(out var value))
+            {
+                yield return value;
+            }
+        }
+    }
+
+    /// <summary>
+    /// Flattens an asynchronous sequence of <see cref="Result{T, TErr}"/> into a sequence containing all error values.
+    /// Ok results are discarded.
+    /// </summary>
+    /// <param name="self">The asynchronsou sequence of results.</param>
+    /// <param name="ct">A CancellationToken that will interrupt async iteration.</param>
+    /// <returns>A flattened sequence of values.</returns>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="self"/> is null.</exception>
+    public static async IAsyncEnumerable<TErr> ErrorsAsync<T, TErr>(this IAsyncEnumerable<Result<T, TErr>> self, [EnumeratorCancellation] CancellationToken ct = default)
+        where T : notnull
+    {
+        ThrowIfNull(self);
+
+        await foreach (var result in self.WithCancellation(ct))
+        {
+            if (result.IsErr(out var err) && err is not null)
+            {
+                yield return err;
+            }
+        }
+    }
 }
